+++ src/main/java/org/sheep1500/toyadvertisementbackend/ads_join/mq/event/AdsJoinEvent.java
@@ -3,6 +3,7 @@ package org.sheep1500.toyadvertisementbackend.ads_join.mq.event;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -11,6 +12,6 @@ public class AdsJoinEvent {
     private String userId;
     private String adId;
     private String adName;
-    private double rewardAmount;
+    private BigDecimal rewardAmount;
     private LocalDateTime joinTime;
 }
